safari-534.53-branch/LayoutTests/platform/qt/fast/events/xsl-onload-expected.txt

CONSOLE MESSAGE: line 3: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>
This is a regression test that checks for a problem where parsing would be stuck true and you'd never get a load event for an XSL-generated page

If the test succeeds, the next paragraph will say "handled load event". If that message is missing, the test failed.

Handled load event.
